feat: implements the token configurations.
