+++ src/main/java/com/example/authorization_service/config/AuthorizationServerConfig.java
@@ -1,10 +1,5 @@
 package com.example.authorization_service.config;
 
-import com.nimbusds.jose.jwk.JWKSet;
-import com.nimbusds.jose.jwk.RSAKey;
-import com.nimbusds.jose.jwk.source.ImmutableJWKSet;
-import com.nimbusds.jose.jwk.source.JWKSource;
-import com.nimbusds.jose.proc.SecurityContext;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.core.annotation.Order;
@@ -13,25 +8,12 @@ import org.springframework.security.config.Customizer;
 import org.springframework.security.config.annotation.web.builders.HttpSecurity;
 import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
 import org.springframework.security.config.core.GrantedAuthorityDefaults;
-import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
-import org.springframework.security.crypto.password.PasswordEncoder;
-import org.springframework.security.oauth2.jwt.JwtDecoder;
-import org.springframework.security.oauth2.server.authorization.config.annotation.web.configuration.OAuth2AuthorizationServerConfiguration;
 import org.springframework.security.oauth2.server.authorization.config.annotation.web.configurers.OAuth2AuthorizationServerConfigurer;
 import org.springframework.security.oauth2.server.authorization.settings.AuthorizationServerSettings;
 import org.springframework.security.oauth2.server.authorization.settings.ClientSettings;
-import org.springframework.security.oauth2.server.authorization.settings.OAuth2TokenFormat;
-import org.springframework.security.oauth2.server.authorization.settings.TokenSettings;
 import org.springframework.security.web.SecurityFilterChain;
 import org.springframework.security.web.authentication.LoginUrlAuthenticationEntryPoint;
 import org.springframework.security.web.util.matcher.MediaTypeRequestMatcher;
-
-import java.security.KeyPair;
-import java.security.KeyPairGenerator;
-import java.security.interfaces.RSAPrivateKey;
-import java.security.interfaces.RSAPublicKey;
-import java.time.Duration;
-import java.util.UUID;
 
 @Configuration
 @EnableWebSecurity
@@ -46,13 +28,15 @@ public class AuthorizationServerConfig {
         http
             .securityMatcher(authorizationConfigure.getEndpointsMatcher())
             .with(authorizationConfigure, (authorizationServer) ->
-                                                                    authorizationServer.oidc(Customizer.withDefaults()))
+                                            authorizationServer.oidc(Customizer.withDefaults()))
             .authorizeHttpRequests(authorize ->
-                                                                    authorize.anyRequest().authenticated())
+                                            authorize.anyRequest().authenticated())
+            .oauth2ResourceServer(oauth ->
+                                            oauth.jwt(Customizer.withDefaults()))
             .exceptionHandling(exceptions ->
-                                                    exceptions.defaultAuthenticationEntryPointFor(
-                                                            new LoginUrlAuthenticationEntryPoint("/login"),
-                                                            new MediaTypeRequestMatcher(MediaType.TEXT_HTML))
+                                            exceptions.defaultAuthenticationEntryPointFor(
+                                                    new LoginUrlAuthenticationEntryPoint("/login"),
+                                                    new MediaTypeRequestMatcher(MediaType.TEXT_HTML))
             );
 
         return http.build();
@@ -64,41 +48,10 @@ public class AuthorizationServerConfig {
 
         http
             .authorizeHttpRequests((authorize) ->
-                                                                        authorize.anyRequest().authenticated())
+                                                authorize.anyRequest().authenticated())
             .formLogin(Customizer.withDefaults());
 
         return http.build();
-    }
-
-    @Bean
-    public JWKSource<SecurityContext> jwkSource() {
-        KeyPair keyPair = generateRsaKey();
-        RSAPublicKey publicKey = (RSAPublicKey) keyPair.getPublic();
-        RSAPrivateKey privateKey = (RSAPrivateKey) keyPair.getPrivate();
-        RSAKey rsaKey = new RSAKey.Builder(publicKey)
-                .privateKey(privateKey)
-                .keyID(UUID.randomUUID().toString())
-                .build();
-        JWKSet jwkSet = new JWKSet(rsaKey);
-        return new ImmutableJWKSet<>(jwkSet);
-    }
-
-    private static KeyPair generateRsaKey() {
-        KeyPair keyPair;
-        try {
-            KeyPairGenerator keyPairGenerator = KeyPairGenerator.getInstance("RSA");
-            keyPairGenerator.initialize(2048);
-            keyPair = keyPairGenerator.generateKeyPair();
-        }
-        catch (Exception ex) {
-            throw new IllegalStateException(ex);
-        }
-        return keyPair;
-    }
-
-    @Bean
-    public JwtDecoder jwtDecoder(JWKSource<SecurityContext> jwkSource) {
-        return OAuth2AuthorizationServerConfiguration.jwtDecoder(jwkSource);
     }
 
     @Bean
@@ -107,23 +60,10 @@ public class AuthorizationServerConfig {
     }
 
     @Bean
-    public TokenSettings tokenSettings() {
-        return TokenSettings.builder()
-                            .accessTokenFormat(OAuth2TokenFormat.SELF_CONTAINED)
-                            .accessTokenTimeToLive(Duration.ofMinutes(60))
-                            .build();
-    }
-
-    @Bean
     public ClientSettings clientSettings() {
         return ClientSettings.builder()
                              .requireAuthorizationConsent(false)
                              .build();
-    }
-
-    @Bean
-    public PasswordEncoder passwordEncoder() {
-        return new BCryptPasswordEncoder(10);
     }
 
     @Bean
