+++ src/main/java/com/example/authorization_service/controllers/ClientController.java
@@ -11,7 +11,7 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
 @RestController
-@RequestMapping("/authorize")
+@RequestMapping("/client")
 public class ClientController {
 
     private final ClientService clientService;
