+++ src/main/java/com/example/authorization_service/repository/CustomRegisteredClientRepository.java
@@ -3,6 +3,7 @@ package com.example.authorization_service.repository;
 import com.example.authorization_service.service.ClientService;
 import org.springframework.security.oauth2.core.AuthorizationGrantType;
 import org.springframework.security.oauth2.core.ClientAuthenticationMethod;
+import org.springframework.security.oauth2.core.oidc.OidcScopes;
 import org.springframework.security.oauth2.server.authorization.client.RegisteredClient;
 import org.springframework.security.oauth2.server.authorization.client.RegisteredClientRepository;
 import org.springframework.security.oauth2.server.authorization.settings.ClientSettings;
