+++ src/main/resources/application.yml
@@ -23,23 +23,3 @@ spring:
     user:
       name: user
       password: password
-  oauth2:
-    authorizationserver:
-      client:
-        oidc-client:
-          registration:
-            client-id: "oidc-client"
-            client-secret: "{noop}secret"
-            client-authentication-methods:
-              - "client_secret_basic"
-            authorization-grant-types:
-              - "authorization_code"
-              - "refresh_token"
-            redirect-uris:
-              - "http://127.0.0.1:8080/login/oauth2/code/oidc-client"
-            post-logout-redirect-uris:
-              - "http://127.0.0.1:8080/"
-            scopes:
-              - "openid"
-              - "profile"
-          require-authorization-consent: true
